# openapi.yml
...
openapi: 3.0.0
info:
  version: 1.0.0
  title: The Owen Wilson Wow API
  license:
    name: MIT
servers:
  - url: /wows
paths:
...
  /ordered/{index}:
    get:
      summary: Ordered "Wow"
      description: Retrieve "wow" results by their index in the chronological order of all results.
      operationId: ordered
      tags:
        - wow
      parameters:
        - name: index
          in: path
          description: Index within results array. Either a single index or a range. Range is between a first index and a second index, inclusive.
          required: true
          examples:
            single:
              value: 0
            range:
              value: 3-7
          schema:
            type: string
      responses:
        "200":
          description: Either a single movie or an array of movies
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Wow"
                  - $ref: "#/components/schemas/Wows"
        "400":
          description: Bad Request
          content:
            text/html:
              schema:
                type: string
              examples:
                response:
                  summary: Invalid index input
                  value: "400 Bad Request: Index should be a number or a range between two numbers"
...
components:
  schemas:
    Wow:
      type: object
      properties:
        movie:
          type: string
          example: Midnight in Paris
        year:
          type: integer
          example: 2011
        release_date:
          type: string
          format: date
          example: 2011-05-20
        director:
          type: string
          example: Woody Allen
        character:
          type: string
          example: Gil Pender
        movie_duration:
          type: string
          example: 01:33:57
        timestamp:
          type: string
          example: 00:58:22
        full_line:
          type: string
          example: Wow.
        current_wow_in_movie:
          type: integer
          example: 5
        total_wows_in_movie:
          type: integer
          example: 10
        poster:
          type: string
          example: https://images.ctfassets.net/bs8ntwkklfua/2ZcfSCe2dlfoVzYMr4b9nK/d566e5ad044dee56645f3bffc7200d64/Midnight_in_Paris_Poster.jpg
        video:
          type: object
          properties:
            1080p:
              type: string
              example: https://videos.ctfassets.net/bs8ntwkklfua/1DSaYjQ8SnL1Imeuxe0eXE/100d3106c8a0bf5e80372e2187daf325/Midnight_in_Paris_Wow_5_1080p.mp4
            720p:
              type: string
              example: https://videos.ctfassets.net/bs8ntwkklfua/eFI1n7voe4CsxkWEWv2q0/36781ed8f50e508e8e1f79c8e65e601a/Midnight_in_Paris_Wow_5_720p.mp4
            480p:
              type: string
              example: https://videos.ctfassets.net/bs8ntwkklfua/6oSaIzfy7k3DJMvvy3j1kN/d77192d7decb3ba9d06a5210a58df1fa/Midnight_in_Paris_Wow_5_480p.mp4
            360p:
              type: string
              example: https://videos.ctfassets.net/bs8ntwkklfua/4LQ4lkhx41XDC6UcMlmjbA/dfd5e3604c85c9fb9d38ea2dfdd02a2f/Midnight_in_Paris_Wow_5_360p.mp4
        audio:
          type: string
          example: https://assets.ctfassets.net/bs8ntwkklfua/2A5G34x8JCQBuYwU4D2l9S/27cc467b3ff5796ff1bbe113a06a6e64/Midnight_in_Paris_Wow_5.mp3
    Wows:
      type: array
      items:
        $ref: "#/components/schemas/Wow"